+++ packages/ui/src/components/TextField/TextField.tsx
@@ -4,6 +4,7 @@ import {
   RegisterOptions,
   FieldError,
   PasswordField,
+  EmailField,
 } from '@redwoodjs/forms'
 import { forwardRef } from 'react'
 
@@ -12,6 +13,7 @@ import './TextField.css'
 interface TextFieldProps {
   name: string
   label: string
+  email?: boolean
   wrapperClassName?: string
   validation?: RegisterOptions
   defaultValue?: string
@@ -25,16 +27,18 @@ const TextField = forwardRef(
     {
       name,
       label,
+      email,
       validation,
       disabled,
       password,
       size,
       defaultValue,
       wrapperClassName,
+      ...rest
     }: TextFieldProps,
     ref: React.ForwardedRef<HTMLInputElement>
   ) => {
-    const Field = password ? PasswordField : RTextField
+    const Field = password ? PasswordField : email ? EmailField : RTextField
     return (
       <div className={wrapperClassName}>
         <Label
@@ -50,6 +54,7 @@ const TextField = forwardRef(
             placeholder=" "
             defaultValue={defaultValue}
             ref={ref}
+            {...rest}
           />
           <span>{label}</span>
         </Label>